requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.36.0
torch==2.4.0
peft==0.7.0
datasets==2.16.0
accelerate==0.24.0
bitsandbytes==0.41.2
numpy==1.24.4
scikit-learn==1.3.2